username passed from gamepage now
