+++ client/components/Chatroom.js
@@ -7,7 +7,7 @@ import {randomNumGenerator} from './utils'
 class Chatroom extends React.Component {
   constructor(props) {
     super(props)
-    this.state = {message: '', username: '', messages: []}
+    this.state = {message: '', username: this.props.username, messages: []}
     this.chatroom = db
       .collection('games')
       .doc(this.props.gamename)
@@ -16,38 +16,14 @@ class Chatroom extends React.Component {
     this.chatroom.onSnapshot(this.listenMessages)
   }
 
-  async componentDidMount() {
-    let userId = ''
-    let username = 'Guest' + randomNumGenerator()
-
-    await firebase.auth().onAuthStateChanged(loggedinUser => {
-      if (loggedinUser) {
-        userId = loggedinUser.uid
-        db
-          .collection('users')
-          .doc(userId)
-          .get()
-          .then(doc => {
-            if (doc.exists) {
-              username = doc.data().username
-              this.setState({username: username})
-            }
-          })
-          .then(() =>
-            this.chatroom.add({
-              username: 'Admin',
-              message: `${username} has entered the room`,
-              createdAt: firebase.firestore.Timestamp.fromDate(new Date())
-            })
-          )
-      }
-    })
+  componentDidMount() {
+    console.log('username', this.props)
     db
       .collection('games')
       .doc('game1')
       .collection('participants')
-      .doc(username)
-      .set({username: username})
+      .doc(this.props.username)
+      .set({username: this.props.username})
 
     window.addEventListener('beforeunload', ev => {
       ev.preventDefault()
@@ -55,11 +31,11 @@ class Chatroom extends React.Component {
         .collection('games')
         .doc('game1')
         .collection('participants')
-        .doc(this.state.username)
+        .doc(this.props.username)
         .delete()
       this.chatroom.add({
         username: 'Admin',
-        message: `${username} has left the room`,
+        message: `${this.props.username} has left the room`,
         createdAt: firebase.firestore.Timestamp.fromDate(new Date())
       })
     })
@@ -76,7 +52,7 @@ class Chatroom extends React.Component {
       .collection('games')
       .doc('game1')
       .collection('participants')
-      .doc(this.state.username)
+      .doc(this.props.username)
       .delete()
       .then(function() {
         console.log('doc successfully deleted')
@@ -87,11 +63,11 @@ class Chatroom extends React.Component {
         .collection('games')
         .doc('game1')
         .collection('participants')
-        .doc(this.state.username)
+        .doc(this.props.username)
         .delete()
       this.chatroom.add({
         username: 'Admin',
-        message: `${this.state.username} has left the room`,
+        message: `${this.props.username} has left the room`,
         createdAt: firebase.firestore.Timestamp.fromDate(new Date())
       })
     })
@@ -108,7 +84,7 @@ class Chatroom extends React.Component {
     if (this.state.username && this.state.message) {
       this.chatroom
         .add({
-          username: this.state.username,
+          username: this.props.username,
           message: this.state.message,
           createdAt: firebase.firestore.Timestamp.fromDate(new Date())
         })
