+++ client/components/Gamepage.js
@@ -1,19 +1,51 @@
 import React from 'react'
 import {Chatroom, CanvasComponent, Test} from './index'
+import firebase from 'firebase'
+import {randomNumGenerator} from './utils'
+import {app, db, config} from '../../firebase-server/firebase'
 
 class Gamepage extends React.Component {
   constructor(props) {
     super(props)
+    this.state = {
+      username: ''
+    }
+  }
+
+  async componentDidMount() {
+    let userId = ''
+    let username = 'Guest' + randomNumGenerator()
+
+    await firebase.auth().onAuthStateChanged(loggedinUser => {
+      if (loggedinUser) {
+        userId = loggedinUser.uid
+        db
+          .collection('users')
+          .doc(userId)
+          .get()
+          .then(doc => {
+            if (doc.exists) {
+              username = doc.data().username
+            }
+          })
+          .then(() => this.setState({username: username}))
+      }
+    })
   }
 
   render() {
-    return (
+    return this.state.username ? (
       <div id="gamepage">
         <CanvasComponent gamename={this.props.match.params.gamename} />
-        <Chatroom gamename={this.props.match.params.gamename} />
+        <Chatroom
+          gamename={this.props.match.params.gamename}
+          username={this.state.username}
+        />
 
         <Test />
       </div>
+    ) : (
+      <div />
     )
   }
 }
